requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
SQLAlchemy==2.0.22
gunicorn==21.2.0
email-validator==2.1.0
python-dateutil==2.8.2 